#1701: extract interface types into types_interfaces.go
Moves the physical/logical interface type definitions verbatim out of
types.go into types_interfaces.go (same package config):

  InterfacesConfig, InterfaceConfig, AggregatedEtherOptions,
  InterfaceUnit, VRRPGroup, DHCPv6ClientConfig, DHCPInetOptions.

The interface-name resolution helpers (LinuxIfName, ResolveReth,
ResolveKernelIfName, DHCPLeaseKey, TunnelNameMap) stay in types.go with
the Config root struct because they are cross-domain resolvers, not pure
interface data types. Pure byte-identical motion; consumers untouched.
Verified: go build ./... clean, go vet clean, config tests pass.
