+++ pkg/config/types.go
@@ -1476,102 +1476,6 @@ type AddressSet struct {
 	AddressSets []string // references to other AddressSet names (nested)
 }
 
-// InterfacesConfig holds interface configuration.
-type InterfacesConfig struct {
-	Interfaces map[string]*InterfaceConfig
-}
-
-// InterfaceConfig represents a network interface.
-type InterfaceConfig struct {
-	Name                string
-	Description         string                  // free-text interface description
-	MTU                 int                     // interface-level MTU (overridden by unit MTU)
-	Speed               string                  // interface speed (e.g. "1g", "10g", "auto")
-	Duplex              string                  // "full", "half", "auto"
-	VlanTagging         bool                    // 802.1Q trunk mode
-	FlexibleVlanTagging bool                    // flexible 802.1Q VLAN tagging (QinQ)
-	Encapsulation       string                  // physical link-layer encapsulation (e.g. "flexible-ethernet-services")
-	Bandwidth           uint64                  // interface bandwidth in bits per second
-	Disable             bool                    // administratively disabled
-	RedundantParent     string                  // gigether-options redundant-parent (HA)
-	LAGParent           string                  // gigether-options 802.3ad <ae-name> (LAG member binding)
-	RedundancyGroup     int                     // redundant-ether-options redundancy-group (0 = none)
-	FabricMembers       []string                // fabric-options member-interfaces
-	LocalFabricMember   string                  // resolved local member for this node (vSRX fabric-options mode)
-	BondMode            string                  // bond mode: "active-backup" for fabric, "802.3ad" for ae
-	AggregatedEtherOpts *AggregatedEtherOptions // ae interface options (LACP, etc.)
-	Units               map[int]*InterfaceUnit
-	Tunnel              *TunnelConfig // non-nil for tunnel interfaces (gre0, etc.)
-}
-
-// AggregatedEtherOptions defines LAG/ae interface parameters.
-type AggregatedEtherOptions struct {
-	LACPActive   bool   // LACP active mode
-	LACPPassive  bool   // LACP passive mode
-	LACPPeriodic string // LACP periodic timer: "fast" or "slow"
-	LinkSpeed    string // required member link speed (e.g. "1g", "10g")
-	MinimumLinks int    // minimum active member links before bundle goes down
-}
-
-// InterfaceUnit represents a logical unit on an interface.
-type InterfaceUnit struct {
-	Number           int
-	Description      string           // free-text unit description
-	VlanID           int              // 0 = native/untagged, >0 = 802.1Q tagged
-	InnerVlanID      int              // inner VLAN tag for QinQ (flexible-vlan-tagging)
-	PointToPoint     bool             // point-to-point link (for tunnels)
-	Addresses        []string         // CIDR notation
-	PrimaryAddress   string           // address marked as primary
-	PreferredAddress string           // address marked as preferred
-	MTU              int              // family-level MTU (0 = default)
-	DHCP             bool             // family inet { dhcp; }
-	DHCPOptions      *DHCPInetOptions // dhcp sub-options (lease-time, etc.)
-	DHCPv6           bool             // family inet6 { dhcpv6; }
-	DHCPv6Client     *DHCPv6ClientConfig
-	DADDisable       bool                  // family inet6 { dad-disable; }
-	SamplingInput    bool                  // family inet/inet6 { sampling { input; } }
-	SamplingOutput   bool                  // family inet/inet6 { sampling { output; } }
-	FilterInputV4    string                // family inet { filter { input NAME; } }
-	FilterOutputV4   string                // family inet { filter { output NAME; } }
-	FilterInputV6    string                // family inet6 { filter { input NAME; } }
-	FilterOutputV6   string                // family inet6 { filter { output NAME; } }
-	VRRPGroups       map[string]*VRRPGroup // keyed by address (CIDR), each address can have VRRP groups
-	Tunnel           *TunnelConfig         // per-unit tunnel config (for multi-unit GRE/IPIP)
-}
-
-// VRRPGroup defines a VRRP (Virtual Router Redundancy Protocol) group.
-type VRRPGroup struct {
-	ID                 int
-	VirtualAddresses   []string // virtual IP addresses
-	Priority           int      // 1-255, default 100
-	Preempt            bool
-	AcceptData         bool
-	AdvertiseInterval  int    // seconds, default 1
-	AuthType           string // "md5" or ""
-	AuthKey            string
-	TrackInterface     string // lower priority if interface is down
-	TrackPriorityDelta int    // how much to lower priority
-}
-
-// DHCPv6ClientConfig holds DHCPv6 client options (dhcpv6-client stanza).
-type DHCPv6ClientConfig struct {
-	DUIDType                   string   // "duid-ll" or "duid-llt"
-	ClientType                 string   // "stateful" or "stateless"
-	ClientIATypes              []string // "ia-pd", "ia-na"
-	PrefixDelegatingPrefixLen  int      // preferred-prefix-length (0 = not set)
-	PrefixDelegatingSubPrefLen int      // sub-prefix-length (0 = not set)
-	ReqOptions                 []string // dns-server, domain-name, etc.
-	UpdateRAInterface          string   // update-router-advertisement interface
-}
-
-// DHCPInetOptions holds DHCPv4 client options for family inet dhcp stanza.
-type DHCPInetOptions struct {
-	LeaseTime              int  // seconds (0 = default)
-	RetransmissionAttempt  int  // number of retransmission attempts
-	RetransmissionInterval int  // seconds between retransmissions
-	ForceDiscover          bool // always start with DISCOVER
-}
-
 // ApplicationsConfig holds application definitions.
 type ApplicationsConfig struct {
 	Applications    map[string]*Application
